+++ ChangeLog.md
@@ -11,3 +11,4 @@
 * Added new XD7StoreFrontFilterKeyword resource
 * Added new XD7StoreFrontOptimalGateway resource
 * Added new XD7StoreFrontRegisterStoreGateway resource
+* Added new XD7StoreFrontSessionStateTimeout resource